+++ services/roma/src/main/java/com/huaweicloud/sdk/roma/v2/model/ListStatisticsApiResponse.java
@@ -20,12 +20,12 @@ public class ListStatisticsApiResponse extends SdkResponse {
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "start_time")
 
-    private Integer startTime;
+    private Long startTime;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "end_time")
 
-    private Integer endTime;
+    private Long endTime;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "list")
@@ -48,7 +48,7 @@ public class ListStatisticsApiResponse extends SdkResponse {
         this.code = code;
     }
 
-    public ListStatisticsApiResponse withStartTime(Integer startTime) {
+    public ListStatisticsApiResponse withStartTime(Long startTime) {
         this.startTime = startTime;
         return this;
     }
@@ -56,15 +56,15 @@ public class ListStatisticsApiResponse extends SdkResponse {
     /** 开始时间的UTC的时间戳
      * 
      * @return startTime */
-    public Integer getStartTime() {
+    public Long getStartTime() {
         return startTime;
     }
 
-    public void setStartTime(Integer startTime) {
+    public void setStartTime(Long startTime) {
         this.startTime = startTime;
     }
 
-    public ListStatisticsApiResponse withEndTime(Integer endTime) {
+    public ListStatisticsApiResponse withEndTime(Long endTime) {
         this.endTime = endTime;
         return this;
     }
@@ -72,11 +72,11 @@ public class ListStatisticsApiResponse extends SdkResponse {
     /** 截止时间的UTC的时间戳
      * 
      * @return endTime */
-    public Integer getEndTime() {
+    public Long getEndTime() {
         return endTime;
     }
 
-    public void setEndTime(Integer endTime) {
+    public void setEndTime(Long endTime) {
         this.endTime = endTime;
     }
